# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "سیستم مدیریت نوبت‌دهی کلینیک دندانپزشکی",
    "description": "API کامل و حرفه‌ای برای مدیریت نوبت‌دهی دو کلینیک دندانپزشکی\n\n⚠️ نکته مهم CSRF Protection:\n1. ابتدا از /auth/csrf-token یک token دریافت کنید\n2. در header درخواست‌های POST/PATCH/DELETE، X-CSRF-Token را با مقدار token ارسال کنید",
    "version": "1.4.0",
    "contact": {
      "name": "Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Development server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Alternative port"
    }
  ],
...
  "paths": {
...
    "/categories/articles": {
      "get": {
        "tags": ["Categories"],
        "summary": "لیست دسته‌بندی‌های مقالات",
        "operationId": "getArticleCategories",
        "parameters": [
          {
            "name": "published",
            "in": "query",
            "schema": { "type": "string", "enum": ["true", "false"] },
            "description": "فقط دسته‌بندی‌های منتشر شده"
          }
        ],
        "responses": {
          "200": {
            "description": "لیست دسته‌بندی‌های مقالات",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "categories": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/ArticleCategory" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "ایجاد دسته‌بندی مقاله (مدیر/منشی)",
        "operationId": "createArticleCategory",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "example": "سلامت دهان و دندان" },
                  "description": { "type": "string", "example": "مقالات مربوط به سلامت دهان و دندان" },
                  "order": { "type": "integer", "example": 1 },
                  "published": { "type": "boolean", "example": true }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "دسته‌بندی ایجاد شد" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/categories/articles/{identifier}": {
      "get": {
        "tags": ["Categories"],
        "summary": "دریافت دسته‌بندی مقاله (با slug یا ID)",
        "operationId": "getArticleCategory",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "دسته‌بندی مقاله" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/categories/articles/{id}": {
      "patch": {
        "tags": ["Categories"],
        "summary": "به‌روزرسانی دسته‌بندی مقاله (مدیر/منشی)",
        "operationId": "updateArticleCategory",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "order": { "type": "integer" },
                  "published": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "دسته‌بندی به‌روزرسانی شد" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "delete": {
        "tags": ["Categories"],
        "summary": "حذف دسته‌بندی مقاله (مدیر/منشی)",
        "operationId": "deleteArticleCategory",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": {
          "200": { "description": "دسته‌بندی حذف شد" },
          "400": { "description": "دسته‌بندی دارای مقاله است" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/categories/services": {
      "get": {
        "tags": ["Categories"],
        "summary": "لیست دسته‌بندی‌های خدمات",
        "operationId": "getServiceCategories",
        "parameters": [
          {
            "name": "published",
            "in": "query",
            "schema": { "type": "string", "enum": ["true", "false"] },
            "description": "فقط دسته‌بندی‌های منتشر شده"
          }
        ],
        "responses": {
          "200": {
            "description": "لیست دسته‌بندی‌های خدمات",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "categories": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/ServiceCategory" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "ایجاد دسته‌بندی خدمت (مدیر/منشی)",
        "operationId": "createServiceCategory",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "example": "جراحی" },
                  "description": { "type": "string", "example": "خدمات جراحی دندانپزشکی" },
                  "order": { "type": "integer", "example": 1 },
                  "published": { "type": "boolean", "example": true }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "دسته‌بندی ایجاد شد" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/categories/services/{identifier}": {
      "get": {
        "tags": ["Categories"],
        "summary": "دریافت دسته‌بندی خدمت (با slug یا ID)",
        "operationId": "getServiceCategory",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "دسته‌بندی خدمت" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/categories/services/{id}": {
      "patch": {
        "tags": ["Categories"],
        "summary": "به‌روزرسانی دسته‌بندی خدمت (مدیر/منشی)",
        "operationId": "updateServiceCategory",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "order": { "type": "integer" },
                  "published": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "دسته‌بندی به‌روزرسانی شد" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "delete": {
        "tags": ["Categories"],
        "summary": "حذف دسته‌بندی خدمت (مدیر/منشی)",
        "operationId": "deleteServiceCategory",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": {
          "200": { "description": "دسته‌بندی حذف شد" },
          "400": { "description": "دسته‌بندی دارای خدمت است" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArticleCategory": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "order": { "type": "integer" },
          "published": { "type": "boolean" },
          "_count": {
            "type": "object",
            "properties": {
              "articles": { "type": "integer" }
            }
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ServiceCategory": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "order": { "type": "integer" },
          "published": { "type": "boolean" },
          "_count": {
            "type": "object",
            "properties": {
              "services": { "type": "integer" }
            }
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string" },
          "errors": { "type": "array", "items": { "type": "string" } }
        }
      }
    },
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "شناسه یکتای رکورد"
      },
...
    },
    "responses": {
...
      "Forbidden": {
        "description": "دسترسی ممنوع - نیاز به نقش مناسب",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "رکورد یافت نشد",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dental.sid",
        "description": "Session-based authentication using cookies"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "CSRF token برای حفاظت در برابر حملات CSRF. ابتدا /auth/csrf-token را صدا کنید و token دریافتی را در این header ارسال کنید"
      }
    }
  }
}